requirements.txt
insightface
mediapipe
numpy
firebase-admin
fastapi
uvicorn[standard]
opencv-python
streamlit
scipy
pydantic
python-multipart
pytest
pytest-asyncio
aiofiles
python-jose[cryptography]
passlib[bcrypt]
celery
apscheduler
scikit-image
onnxruntime
python-dotenv
